chore: tests in ts
